Activate pascalprecht.translate
Going to try out simply adding the translate directive to HTML tags for
angular, so that this can be a simple additive step for languages that
aren't en-US. If angular-translate doesn't have a translation, it falls
back to whatever's in the tag.

The alternative is to go with <tag>OBVIOUS_UNTRANSLATED</tag> so that
the notice when new stuff shows up. Discuss this with Nick/Luigi.