Changes new columns to imperial canonical
NREL has opted for keeping the canonical database units as imperial, and
so Canadian use will simply be translation at the UI layer. Converts the
pint columns to simply be `_pint` suffixed instead of `_si` since our
swapping-in strategy will be the same. The data migration is now a
straight copy of float->float from Postgres's point of view.