Adds pint columns and a `pint` data type
No actual idea yet if this is the right way to go but we need to
signal somehow to the mapper that these values have meaning and must be
treated specially (ie. like how we convert floats).

The pint columns are evident in the UI as either a "Gross Floor Area*"
or "Gross Floor Area Pint" depending on whether the col shows with a
de-underscoring or by actually using the display name.