Round-trips incoming pint data through pint
Currently hard-coded to incoming m**2 and outgoing ft**2 (implicitly by
the model field definition). Adds a separate pint_cleaner that
(somehow?) can convert incoming data to a Quantity object with the right
units. Adds a new serializer to just return the Quantity object as a
string until we figure out what to do with this in the UI.