Adds a pint_unit column to the Column model
Seems like the cleaner gets constructed from the Column, not the
ColumnMappings per se, so let's try including the pint unit on here
instead and see what happens. The data "type" is more like a
parameterized type representable as a tuple ('pint_object', 'm**2') than
a simple 'string' or 'float', so let's see if we can dispatch on that.