Teaches the cleaner about units
Sneaking units into the cleaner as a poor-man's parameterized type ...
eg. it's not enough to say it's type `Pint`, or type `EUI`, but has to
be something like `(Pint, 'GJ/m**2/year')` where the Pint format string
is the type parameter. This'll give us a few benefits:

- Relatively easy dispatch within the cleaner... the actual units has to
come from *somewhere* and this seems the most direct / debuggable.

- We'll get automatic conversion on the way into the db courtesy of the
  django-pint custom field.

- We'll see a `DimensionalityError` thrown from Pint if the unit string
  isn't appropriate (eg. feed something tagged 'm**2' into a field with
  base units `kBtu/ft**2/year`.

It's not great that this bleeds the implementation detail of the pint
unit string upwards so far and propagates "stringly" typed objects
(which always suck), so future work is to define the conversions we're
interested in as named consts someplace. But until we actually get unit
conversions in and out working this is the least tedious option.