Makes dq units submit as pint units
Should make these constants; filing away as a future refactoring when I
actually get it working.

There'll be a migration to apply to old rules so they correctly
populate using the new pint units column. For all that, it's possible we
just re-use the existing `seed_rule.units` column in place since it's
pretty unambiguous as a translation.