Adds `units_pint` to column model
A given mapping has to store not just its raw<->mapped name match but
also what units to use for it. Eg. the raw names "Source EUI
(GJ/m2/year)" and "Source EUI (kBtu/ft2/year)" could both map to
`source_eui_pint`, so logically this has to be a property of the raw
column. (Mapped columns will also get the column since there's no
distinction at the storage layer; we'll just not use it in the code).