Clarifies names in mapper
The canonical names for these concepts in the system seem to be "raw"
and "mapped" ... `value` and `field_name` are a bit abstract and I'd
been losing track of what each signified. `raw_field` wasn't too bad,
but it's better (for me, anyway) to understand them as being a column
name ... it's not a value.