Reworks `_build_cleaner` to add pint unit data to cleaners
Adds drop-in replacement for old `_build_cleaner` minus the
unused-in-2.0 units stuff. Reverses the build order: now *starts* with
the default types from the giant array in the columns file, and
supplements it with any pint types we can find on this org's columns.

The result is a cleaner that has access to unit data, with columns
possibly having the type `('quantity', 'some-pint-string')` instead of
just a string like `'date'` or `'number'`. But, we'll have to dispatch
the cleaning on the raw name to get access to it.